pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-contract-validator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "rich>=13.7",
    "click>=8.1",
]

[project.scripts]
validate-contract = "data_contract_validator.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.3",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[tool.hatch.build.targets.wheel]
packages = ["src/data_contract_validator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
